Refactor adapter package structure by reorganizing gateway, HTTP client, and persistence layers
