+++ internal/adapter/gateway/parts_supply_gateway.go
