+++ internal/adapter/gateway/vehicle_gateway.go
