+++ internal/adapter/http/entity_api/endpoints.go
