+++ internal/adapter/http/entity_api/vehicles_repository.go
